Move `SchedulerConfig` from `config/__init__.py` to `config/scheduler.py`
